0.16.0

+++ Source/Project/Documents/Object.cpp
@@ -1,5 +1,5 @@
 //---------------------------------------------------------------------------
-#include "AgdStudio.pch.h"
+#include "AGD Studio.pch.h"
 //---------------------------------------------------------------------------
 #include "Object.h"
 //---------------------------------------------------------------------------
@@ -47,8 +47,8 @@ void __fastcall ObjectDocument::DoSaveExtra()
     Write("RoomIndex", m_RoomIndex);
     Write("State", m_State);
     Push("Position");
-        Write("X", (int)m_Position.X);
-        Write("Y", (int)m_Position.Y);
+        Write("X", m_Position.X);
+        Write("Y", m_Position.Y);
     Pop();
 }
 //---------------------------------------------------------------------------
